+++ solidity/contracts/token/extensions/HypFiatTokenCollateral.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
 pragma solidity >=0.8.0;
 
 import {IFiatToken} from "../interfaces/IFiatToken.sol";