+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hyperlane-xyz/infra
 
+## 3.11.1
+
+### Patch Changes
+
+- Updated dependencies [c900da187]
+  - @hyperlane-xyz/sdk@3.11.1
+  - @hyperlane-xyz/helloworld@3.11.1
+  - @hyperlane-xyz/utils@3.11.1
+
 ## 3.11.0
 
 ### Minor Changes